refactor: App에 있던 컴포넌트 파일로 분리, 각 컴포넌트에 필요한 context만 사용하도록 수정
