+++ src/components/Layout.tsx
@@ -1,0 +1,21 @@
+import { ReactNode } from "react";
+import Header from "./Header";
+import { useTheme } from "@/context/ThemeContext";
+
+interface LayoutProps {
+  children: ReactNode;
+}
+
+const Layout: React.FC<LayoutProps> = ({ children }) => {
+  const { theme } = useTheme();
+  return (
+    <div
+      className={`min-h-screen ${theme === "light" ? "bg-gray-100" : "bg-gray-900 text-white"}`}
+    >
+      <Header />
+      {children}
+    </div>
+  );
+};
+
+export default Layout;
